pyproject.toml
[project]
name = "asf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "channels>=4.3.1",
    "cryptography>=46.0.3",
    "django>=5.2.7",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "dnslib>=0.9.26",
    "dnspython>=2.8.0",
    "jinja2>=3.1.6",
    "langchain-chroma",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-redis>=0.1.2",
    "lxml>=6.0.2",
    "mcp[cli]>=1.19.0",
    "openpyxl>=3.1.5",
    "pluginbase>=1.0.1",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.3",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "rsa>=4.9.1",
    "sentence-transformers>=5.1.2",
    "thehive4py>=2.0.3",
    "tldextract>=5.3.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.38.0",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
